refactor(tool): Move modifier map to utils module
Currently, `HashMap` based modifiers are only used in the verify
feature, but it can be utilized in the strip feature. Before consuming
them in the strip feature, this commit moves them to utils module as
general data structures.
