+++ src/cpu-template-helper/src/verify/mod.rs
@@ -2,9 +2,8 @@
 // SPDX-License-Identifier: Apache-2.0
 
 use std::collections::HashMap;
-use std::fmt::{Binary, Display};
-use std::hash::Hash;
-use std::ops::{BitAnd, Shl};
+
+use crate::utils::{ModifierMapKey, ModifierMapValue};
 
 #[cfg(target_arch = "aarch64")]
 mod aarch64;
@@ -22,64 +21,6 @@ pub enum Error {
     KeyNotFound(String),
     #[error("Value for {0} mismatched.\n{1}")]
     ValueMismatched(String, String),
-}
-
-/// Trait for key of `HashMap`-based modifier.
-///
-/// This is a wrapper trait of some traits required for a key of `HashMap` modifier.
-pub trait ModifierMapKey: Eq + PartialEq + Hash + Display {}
-
-/// Trait for value of `HashMap`-based modifier.
-pub trait ModifierMapValue {
-    // The data size of `Self::Type` varies depending on the target modifier.
-    // * x86_64 CPUID: `u32`
-    // * x86_64 MSR: `u64`
-    // * aarch64 registers: `u128`
-    //
-    // These trait bounds are required for the following reasons:
-    // * `PartialEq + Eq`: To compare `Self::Type` values (like `filter()` and `value()`).
-    // * `BitAnd<Output = Self::Type>`: To use AND operation (like `filter() & value()`).
-    // * `Binary`: To display in a bitwise format.
-    // * `From<bool> + Shl<usize, Output = Self::Type>`: To construct bit masks in
-    //   `to_diff_string()`.
-    type Type: PartialEq
-        + Eq
-        + Copy
-        + BitAnd<Output = Self::Type>
-        + Binary
-        + From<bool>
-        + Shl<usize, Output = Self::Type>;
-
-    // Return `filter` of arch-specific `RegisterValueFilter` in the size for the target.
-    fn filter(&self) -> Self::Type;
-
-    // Return `value` of arch-specific `RegisterValueFilter` in the size for the target.
-    fn value(&self) -> Self::Type;
-
-    // Generate a string to display difference of filtered values between CPU template and guest
-    // CPU config.
-    #[rustfmt::skip]
-    fn to_diff_string(template: Self::Type, config: Self::Type) -> String {
-        let nbits = std::mem::size_of::<Self::Type>() * 8;
-
-        let mut diff = String::new();
-        for i in (0..nbits).rev() {
-            let mask = Self::Type::from(true) << i;
-            let template_bit = template & mask;
-            let config_bit = config & mask;
-            diff.push(match template_bit == config_bit {
-                true => ' ',
-                false => '^',
-            });
-        }
-
-        format!(
-            "* CPU template     : 0b{template:0width$b}\n\
-             * CPU configuration: 0b{config:0width$b}\n\
-             * Diff             :   {diff}",
-            width = nbits,
-        )
-    }
 }
 
 /// Verify that the given CPU template is applied as intended.
@@ -114,6 +55,8 @@ where
 
 #[cfg(test)]
 mod tests {
+    use std::fmt::Display;
+
     use super::*;
 
     #[derive(PartialEq, Eq, Hash)]
